// lib/enums.ts
export const enum MessageImportance {
    IMPORTANCE_LOW = 0,
    IMPORTANCE_NORMAL = 1,
    IMPORTANCE_HIGH = 2,
}

export const enum MessageIconIndex {
    NewMail = 0x00000000,
    Post = 0x00000001,
    Other = 0x00000003,
    ReadMail = 0x00000100,
    UnreadMail = 0x00000101,
    SubmittedMail = 0x00000102,
    UnsentMail = 0x00000103,
    ReceiptMail = 0x00000104,
    RepliedMail = 0x00000105,
    ForwardedMail = 0x00000106,
    RemoteMail = 0x00000107,
    DeliveryReceipt = 0x00000108,
    ReadReceipt = 0x00000109,
    NondeliveryReport = 0x0000010a,
    NonReadReceipt = 0x0000010b,
    RecallSMail = 0x0000010c,
    RecallFMail = 0x0000010d,
    TrackingMail = 0x0000010e,
    OutOfOfficeMail = 0x0000011b,
    RecallMail = 0x0000011c,
    TrackedMail = 0x00000130,
    Contact = 0x00000200,
    DistributionList = 0x00000202,
    StickyNoteBlue = 0x00000300,
    StickyNoteGreen = 0x00000301,
    StickyNotePink = 0x00000302,
    StickyNoteYellow = 0x00000303,
    StickyNoteWhite = 0x00000304,
    SingleInstanceAppointment = 0x00000400,
    RecurringAppointment = 0x00000401,
    SingleInstanceMeeting = 0x00000402,
    RecurringMeeting = 0x00000403,
    MeetingRequest = 0x00000404,
    Accept = 0x00000405,
    Decline = 0x00000406,
    Tentatively = 0x00000407,
    Cancellation = 0x00000408,
    InformationalUpdate = 0x00000409,
    TaskTask = 0x00000500,
    UnassignedRecurringTask = 0x00000501,
    AssigneesTask = 0x00000502,
    AssignersTask = 0x00000503,
    TaskRequest = 0x00000504,
    TaskAcceptance = 0x00000505,
    TaskRejection = 0x00000506,
    JournalConversation = 0x00000601,
    JournalEmailMessage = 0x00000602,
    JournalMeetingRequest = 0x00000603,
    JournalMeetingResponse = 0x00000604,
    JournalTaskRequest = 0x00000606,
    JournalTaskResponse = 0x00000607,
    JournalNote = 0x00000608,
    JournalFax = 0x00000609,
    JournalPhoneCall = 0x0000060a,
    JournalLetter = 0x0000060c,
    JournalMicrosoftOfficeWord = 0x0000060d,
    JournalMicrosoftOfficeExcel = 0x0000060e,
    JournalMicrosoftOfficePowerPoint = 0x0000060f,
    JournalMicrosoftOfficeAccess = 0x00000610,
    JournalDocument = 0x00000612,
    JournalMeeting = 0x00000613,
    JournalMeetingCancellation = 0x00000614,
    JournalRemoteSession = 0x00000615,
}

export type AddressType = "" | "EX" | "SMTP" | "FAX" | "MHS" | "PROFS" | "X400"

export const enum MapiObjectType {
    // Address book container object
    MAPI_ABCONT = 4,
    // Address book object
    MAPI_ADDRBOOK = 2,
    // Message attachment object
    MAPI_ATTACH = 7,
    // Distribution list object
    MAPI_DISTLIST = 8,
    // Folder object
    MAPI_FOLDER = 3,
    // Form object
    MAPI_FORMINFO = 12,
    // Messaging user object
    MAPI_MAILUSER = 6,
    // Message object
    MAPI_MESSAGE = 5,
    // Profile section object
    MAPI_PROFSECT = 9,
    // Session object
    MAPI_SESSION = 11,
    // Status object
    MAPI_STATUS = 10,
    // Message store object
    MAPI_STORE = 1,
}
// ...
export const enum RecipientType {
    // The recipient is the message originator
    Originator = 0x0000,
    // The recipient is a primary (To) recipient. Clients are required to handle primary recipients. All other types are optional.
    To = 0x0001,
    // The recipient is a carbon copy (CC) recipient, a recipient that receives a message in addition to the primary recipients.
    Cc = 0x0002,
    // The recipient is a blind carbon copy (BCC) recipient. Primary and carbon copy recipients are unaware of the existence of BCC recipients.
    Bcc = 0x0003,
    // The recipient is a resource (e.g. a room)
    Resource = 0x0004,
    // The recipient is a room (uses PR_RECIPIENT_TYPE_EXE) needs Exchange 2007 or higher
    Room = 0x0007,
}
// ...
export const enum PropertyFlag {
    // If this flag is set for a property, that property MUST NOT be deleted from the .msg file
    // (irrespective of which storage it is contained in) and implementations MUST return an error
    // if any attempt is made to do so. This flag is set in circumstances where the implementation
    // depends on that property always being present in the .msg file once it is written there.
    PROPATTR_MANDATORY = 0x00000001,
    // If this flag is not set on a property, that property MUST NOT be read from the .msg file
    // and implementations MUST return an error if any attempt is made to read it. This flag is
    // set on all properties unless there is an implementation-specific reason to prevent a property
    // from being read from the .msg file.
    PROPATTR_READABLE = 0x00000002,
    // If this flag is not set on a property, that property MUST NOT be modified or deleted and
    // implementations MUST return an error if any attempt is made to do so. This flag is set in
    // circumstances where the implementation depends on the properties being writable.
    PROPATTR_WRITABLE = 0x00000004,
}

export const enum MessageClass {
    Unknown = "",
    IPM_Note = "IPM.Note",
    IPM_Note_SMIME = "IPM.Note.SMIME",
    IPM_Note_SMIME_MultipartSigned = "IPM.Note.SMIME.MultipartSigned",
    IPM_Note_Receipt_SMIME = "IPM.Note.Receipt.SMIME",
    IPM_Post = "IPM.Post",
    IPM_Octel_Voice = "IPM.Octel.Voice",
    IPM_Voicenotes = "IPM.Voicenotes",
    IPM_Sharing = "IPM.Sharing",
    REPORT_IPM_NOTE_NDR = "REPORT.IPM.NOTE.NDR",
    REPORT_IPM_NOTE_DR = "REPORT.IPM.NOTE.DR",
    REPORT_IPM_NOTE_DELAYED = "REPORT.IPM.NOTE.DELAYED",
    REPORT_IPM_NOTE_IPNRN = "*REPORT.IPM.NOTE.IPNRN",
    REPORT_IPM_NOTE_IPNNRN = "*REPORT.IPM.NOTE.IPNNRN",
    REPORT_IPM_SCHEDULE_MEETING_REQUEST_NDR = "REPORT.IPM.SCHEDULE. MEETING.REQUEST.NDR",
    REPORT_IPM_SCHEDULE_MEETING_RESP_POS_NDR = "REPORT.IPM.SCHEDULE.MEETING.RESP.POS.NDR",
    REPORT_IPM_SCHEDULE_MEETING_RESP_TENT_NDR = "REPORT.IPM.SCHEDULE.MEETING.RESP.TENT.NDR",
    REPORT_IPM_SCHEDULE_MEETING_CANCELED_NDR = "REPORT.IPM.SCHEDULE.MEETING.CANCELED.NDR",
    REPORT_IPM_NOTE_SMIME_NDR = "REPORT.IPM.NOTE.SMIME.NDR",
    REPORT_IPM_NOTE_SMIME_DR = "*REPORT.IPM.NOTE.SMIME.DR",
    REPORT_IPM_NOTE_SMIME_MULTIPARTSIGNED_NDR = "*REPORT.IPM.NOTE.SMIME.MULTIPARTSIGNED.NDR",
    REPORT_IPM_NOTE_SMIME_MULTIPARTSIGNED_DR = "*REPORT.IPM.NOTE.SMIME.MULTIPARTSIGNED.DR",
    IPM_Appointment = "IPM.Appointment",
    IPM_Task = "IPM.Task",
}


export enum MessageEditorFormat {
    // The format for the editor to use is unknown.
    EDITOR_FORMAT_DONTKNOW = 0x00000000,
    // The editor should display the message in plain text format.
    EDITOR_FORMAT_PLAINTEXT = 0x00000001,
    // The editor should display the message in HTML format.
    EDITOR_FORMAT_HTML = 0x00000002,
    // The editor should display the message in Rich Text Format.
    EDITOR_FORMAT_RTF = 0x00000003,
}

export const enum MessageFormat {
    TextOnly = 0,
    HtmlOnly = 1,
    TextAndHtml = 2,
}

export const enum MessagePriority {
    PRIO_NONURGENT = 0,
    PRIO_NORMAL = 1,
    PRIO_URGENT = 2,
}
// ...
export const StoreSupportMaskConst =
    StoreSupportMask.STORE_ATTACH_OK | // | StoreSupportMask.STORE_CATEGORIZE_OK
    StoreSupportMask.STORE_CREATE_OK | //StoreSupportMask.STORE_ENTRYID_UNIQUE
    StoreSupportMask.STORE_MODIFY_OK | // | StoreSupportMask.STORE_MV_PROPS_OK
    // | StoreSupportMask.STORE_OLE_OK
    // | StoreSupportMask.STORE_RTF_OK
    StoreSupportMask.STORE_HTML_OK |
    StoreSupportMask.STORE_UNICODE_OK
// ...
export const enum MessageFlags {
    // The message is marked as having been read. This can occur as the result of a call at any time to
    // IMessage==SetReadFlag or IMAPIFolder==SetReadFlags. Clients can also set this flag by calling a message's
    // IMAPIProp==SetProps method before the message has been saved for the first time. This flag is ignored if the
    // ASSOCIATED flag is set.
    MSGFLAG_READ = 0x0001,
    // The outgoing message has not been modified since the first time that it was saved; the incoming message has not
    // been modified since it was delivered.
    MSGFLAG_UNMODIFIED = 0x0002,
    // The message is marked for sending as a result of a call to IMessage==SubmitMessage. Message store providers set
    // this flag; the client has read-only access.
    MSGFLAG_SUBMIT = 0x0004,
    // The message is still being composed. It is saved, but has not been sent. The client or provider has read/write
    // access to this flag until the first IMAPIProp==SaveChanges call and read-only thereafter. If a client doesn't set
    // this flag by the time the message is sent, the message store provider sets it when IMessage==SubmitMessage is
    // called. Typically, this flag is cleared after the message is sent.
    MSGFLAG_UNSENT = 0x0008,
    // The message has at least one attachment. This flag corresponds to the message's PR_HASATTACH (PidTagHasAttachments)
    // property. The client has read-only access to this flag.
    MSGFLAG_HASATTACH = 0x0010,
    // The messaging user sending was the messaging user receiving the message. The client or provider has read/write
    // access to this flag until the first IMAPIProp==SaveChanges call and read-only thereafter. This flag is meant to be
    // set by the transport provider.
    MSGFLAG_FROMME = 0x0020,
    // The message is an associated message of a folder. The client or provider has read-only access to this flag. The
    // READ flag is ignored for associated messages, which do not retain a read/unread state.
    MSGFLAG_ASSOCIATED = 0x040,
    // The message includes a request for a resend operation with a nondelivery report. The client or provider has
    // read/write access to this flag until the first IMAPIProp==SaveChanges call and read-only thereafter.
    MSGFLAG_RESEND = 0x0080,
    // A read report needs to be sent for the message. The client or provider has read-only access to this flag.
    MSGFLAG_NOTIFYREAD = 0x100,
    // A nonread report needs to be sent for the message. The client or provider has read-only access to this flag.
    MSGFLAG_NOTIFYUNREAD = 0x0200,
    // The message has been read at least once. This flag is set or cleared by the server whenever the MSGFLAG_READ flag
    // is set or cleared.
    MSGFLAG_EVERREAD = 0x0400,
    // The incoming message arrived over an X.400 link. It originated either outside the organization or from a source the
    // gateway cannot consider trusted. The client should display an appropriate message to the user. Transport providers
    // set this flag; the client has read-only access.
    MSGFLAG_ORIGIN_X400 = 0x1000,
    // The incoming message arrived over the Internet. It originated either outside the organization or from a source the
    // gateway cannot consider trusted. The client should display an appropriate message to the user. Transport providers
    // set this flag; the client has read-only access.
    MSGFLAG_ORIGIN_INTERNET = 0x2000,
    // The incoming message arrived over an external link other than X.400 or the Internet. It originated either outside
    // the organization or from a source the gateway cannot consider trusted. The client should display an appropriate
    // message to the user. Transport providers set this flag; the client has read-only access.
    MSGFLAG_ORIGIN_MISC_EXT = 0x8000,
}

/**
 * Contains a bitmask of flags indicating the operations that are available to the client for the object.
 * See https=//msdn.microsoft.com/en-us/library/office/cc979218.aspx
 * This property is read-only for the client. It must be a bitwise OR of zero or more values from the following table.
 */
export const enum MapiAccess {
    // write
    MAPI_ACCESS_MODIFY = 0x00000001,
    // read
    MAPI_ACCESS_READ = 0x00000002,
    // delete
    MAPI_ACCESS_DELETE = 0x00000004,
    // Create subfolders in the folder hierarchy
    MAPI_ACCESS_CREATE_HIERARCHY = 0x00000008,
    // Create content messages
    MAPI_ACCESS_CREATE_CONTENTS = 0x00000010,
    // Create associated content messages
    MAPI_ACCESS_CREATE_ASSOCIATED = 0x00000020,
}